% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges_to_nodes.R
\name{edgesToNodes}
\alias{edgesToNodes}
\title{Separate edges and nodes from graph object}
\usage{
edgesToNodes(PCgrph)
}
\arguments{
\item{PCgrph}{Network graph to split}
}
\value{
tbl_graph objects for further analysis
}
\description{
Separate edges and nodes from graph object
}
