% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makewgraph.R
\name{makeWgraph}
\alias{makeWgraph}
\title{Join with w}
\usage{
makeWgraph(regsobj, PClab)
}
\arguments{
\item{regsobj}{Object with regimes}

\item{PClab}{Label for the Principal Component}
}
\value{
Joined data frame
}
\description{
Join with w
}
