% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_shifts.R
\name{regimes_summary}
\alias{regimes_summary}
\title{Summarize shifts and regimes}
\usage{
regimes_summary(model)
}
\arguments{
\item{model}{Model object with shifts}
}
\value{
Tibble with summary statistics.
}
\description{
Summarize shifts and regimes
}
