% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_convergence.R
\name{summarise_convregs}
\alias{summarise_convregs}
\title{Summarize convergent regimes across a list of models}
\usage{
summarise_convregs(convRegList)
}
\arguments{
\item{convRegList}{List of model objects}
}
\value{
Data frame with summarized regimes
}
\description{
Summarize convergent regimes across a list of models
}
