% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_lasso.R
\name{summarize.lasso.convergence.output}
\alias{summarize.lasso.convergence.output}
\title{Summarize lasso convergence output}
\usage{
summarize.lasso.convergence.output(model)
}
\arguments{
\item{model}{l1ou model object}
}
\value{
Data frame with collapsed sets of convergent taxa and regimes
}
\description{
Summarize lasso convergence output
}
\details{
Modified from Slater and Friscia (2019)
}
