% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_reg_components.R
\name{to_regComponents}
\alias{to_regComponents}
\title{Consensus from edge list}
\usage{
to_regComponents(PCsummary, comparison, qtile, wPC, wmtype)
}
\arguments{
\item{PCsummary}{Edge list in data frame format}

\item{comparison}{Relational operator for comparison}

\item{qtile}{Quantile to retain}

\item{wPC}{Label for PC}

\item{wmtype}{Measurement type label}
}
\value{
Data frame with regimes in long format
}
\description{
Consensus from edge list
}
